pyproject.toml
[tool.poetry]
name = "networking-app-backend"
version = "0.1.0"
description = ""
authors = ["Reuben-Stephen-John <reuben.sj2003@gmail.com>"]
readme = "README.md"
packages = [{include = "networking_app_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.1"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.0.0"
django-import-export = "^3.2.0"
djangorestframework-simplejwt = "^5.2.2"
gunicorn = "^21.2.0"
dj-database-url = "^2.0.0"
django-environ = "^0.10.0"
social-auth-app-django = "^5.2.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-taggit = "^4.0.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.1.0"}
shortuuid = "^1.0.11"
django-allauth = "^0.55.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
